modified:   src/model/Courier.java 	modified:   src/model/Plan.java 	modified:   src/xml/XMLdeserializer.java

+++ src/model/Plan.java
@@ -11,7 +11,7 @@ public class Plan {
 		this.warehouse = null;
 	}
 	
-	public void addWarehouse(float intersectionID) {
+	public void addWarehouse(Long intersectionID) {
 		for(Intersection node : this.nodes) {
 			if(node.getId() == intersectionID) {
 				this.warehouse = node;
